remove: The GeneralUpdate.AspNetCore component is no longer maintained
The core of this component is a unified common method, only SignalR push and a version verification function. Subsequent iterations of the function of design will no longer be a company for each gm for example will depend on the user authentication, access to the file server, and each company is different so the following sample code will be handed over to the GeneralLibrary/GeneralUpdate - Samples in maintenance warehouse. No longer as a standalone component.