feat(Client): Add patch upgrade feature and optimize update process
Add PatchEnabled property in GlobalConfigInfo to control the patch upgrade feature

Initialize the PatchEnabled configuration item in GeneralClientBootstrap

Set the default patch upgrade option to false in Program

Decide whether to use PatchMiddleware in WindowsStrategy based on the configuration item

Add the Patch option in UpdateOption to control the patch upgrade feature, with a default value of true. Binary diff patch upgrade is enabled by default.